# docs/openapi.yml
openapi: 3.0.3
info:
  title: LilyOpenCMS API
  description: |
    A comprehensive API for LilyOpenCMS content management system.
    
    ## Features
    - **Content Management**: News, albums, chapters, and media
    - **User Management**: Profiles, authentication, and permissions
    - **Public API**: Multiplatform access for mobile apps and third-party integrations
    - **SEO Management**: Comprehensive SEO tools and analytics
    - **Analytics**: Performance monitoring and user statistics
    
    ## Authentication
    Most endpoints require authentication via session cookies or JWT tokens.
    Public API endpoints (`/api/public/*`) do not require authentication.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Please implement appropriate retry logic.
    
    ## Support
    For API support, please refer to the documentation or contact the development team.
  version: 1.0.0
  contact:
    name: LilyOpenCMS Development Team
    email: support@lilyopencms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}"
    description: Custom server (set scheme/host below to try any deployment)
    variables:
      scheme:
        enum:
          - http
          - https
        default: https
      host:
        default: localhost:5000
        description: Hostname and optional port (e.g., api.example.com or localhost:5000)
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/news:
    get:
      tags:
        - News
      summary: List news articles/stories (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: q
          in: query
          description: Search query
          schema: { type: string }
        - name: category
          in: query
          description: Filter by category id
          schema: { type: integer }
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [newest, oldest, popular, least-popular]
            default: newest
      responses:
        '200':
          description: List of news
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicNewsListResponse'
    post:
      tags:
        - News
      summary: Create a new story/chapter (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                content: { type: string }
                category: { type: integer, nullable: true }
                date: { type: string, format: date-time, nullable: true }
                age_rating: { type: string, nullable: true }
                is_news: { type: boolean, default: false }
                is_main_news: { type: boolean, default: false }
                is_premium: { type: boolean, default: false }
                is_visible: { type: boolean, default: false }
                writer: { type: integer, nullable: true }
                external_source: { type: string, nullable: true }
                image_id: { type: integer, nullable: true }
                prize: { type: number, nullable: true }
                prize_coin_type: { type: string, enum: [achievement, topup], nullable: true }
                album_id: { type: integer, nullable: true }
              required: [title]
      responses:
        '201':
          description: News created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicNewsResponse'
...
  /api/news/{news_id}:
    get:
      tags:
        - News
      summary: Get story/chapter details (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: news_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: News details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicNewsResponse'
        '404': { description: Not found }
    put:
      tags:
        - News
      summary: Update story/chapter (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: news_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsUpdatePayload'
      responses:
        '200':
          description: News updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicNewsResponse'
        '404': { description: Not found }
    delete:
      tags:
        - News
      summary: Delete story/chapter (admin-tier)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: news_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '403': { description: Forbidden }
        '404': { description: Not found }
...
  /api/search/news:
    get:
      tags:
        - News
      summary: Unified news search API (public)
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: integer }
        - name: category_name
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema:
            type: string
            enum: [general, news, articles, utama]
            default: general
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, popular, least-popular]
            default: newest
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicNewsListResponse'
...
  /api/albums:
    get:
      tags:
        - Albums
      summary: List albums (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: q
          in: query
          description: Search query
          schema: { type: string }
        - name: category
          in: query
          description: Filter by category id
          schema: { type: integer }
        - name: status
          in: query
          description: Filter by album status
          schema:
            type: string
            enum: [ongoing, completed, hiatus]
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [newest, oldest, popular, least-popular, most-viewed, least-viewed, highest-rated]
            default: newest
      responses:
        '200':
          description: List of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAlbumsListResponse'
    post:
      tags:
        - Albums
      summary: Create a new album (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                category_id: { type: integer, nullable: true }
                is_premium: { type: boolean, default: false }
                is_completed: { type: boolean, default: false }
                is_hiatus: { type: boolean, default: false }
                image_id: { type: integer, nullable: true }
              required: [title]
      responses:
        '201':
          description: Album created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAlbumResponse'
  /api/albums/{album_id}:
    get:
      tags:
        - Albums
      summary: Get album details (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Album details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAlbumResponse'
        '404': { description: Not found }
    put:
      tags:
        - Albums
      summary: Update album (authenticated)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                category_id: { type: integer, nullable: true }
                is_premium: { type: boolean }
                is_completed: { type: boolean }
                is_hiatus: { type: boolean }
                image_id: { type: integer, nullable: true }
      responses:
        '200':
          description: Album updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAlbumResponse'
        '404': { description: Not found }
    delete:
      tags:
        - Albums
      summary: Delete album (admin-tier)
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '403': { description: Forbidden }
        '404': { description: Not found }
...
  /api/search/albums:
    get:
      tags:
        - Albums
      summary: Search albums across entire database (public)
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, completed, hiatus]
        - name: rating
          in: query
          schema: { type: integer, minimum: 1, maximum: 5 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, popular, least-popular, most-viewed, least-viewed, highest-rated]
            default: newest
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicAlbumsListResponse'

  # =============================================================================
  # USERS MANAGEMENT & ACCOUNT ENDPOINTS
  # =============================================================================
...
  /api/images:
    get:
      tags:
        - Media
      summary: List images with pagination, filters, and secure ownership-based access control
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: visibility
          in: query
          description: Filter by visibility (visible/hidden)
          schema: { type: string }
        - name: all_users
          in: query
          description: When true, applies proper ownership filtering - regular users see own images + admin/suadmin visible images, editors see own + assigned writers' + admin/suadmin visible images, admins see all images
          schema: { type: boolean }
      responses:
        '200': { description: Images list returned }
    post:
      tags:
        - Media
      summary: Upload a new image
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                is_visible: { type: boolean }
      responses:
        '201': { description: Image uploaded }
  /api/images/{image_id}:
    get:
      tags:
        - Media
      summary: Get image details
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Image details returned }
        '404': { description: Not found }
    put:
      tags:
        - Media
      summary: Update image
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
      responses:
        '200': { description: Image updated }
        '404': { description: Not found }
    delete:
      tags:
        - Media
      summary: Delete image
      security:
        - sessionAuth: []
        - bearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication via cookies
    
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external ad access and third-party integrations
security:
  - sessionAuth: []
  - bearerAuth: []